refactor: move NDJSON streaming into client.go and accept dst parameter
Address PR #780 review feedback from @ElyarSadig:

1. GetTaskDocumentsWithContext now takes a dst interface{} pointer-to-slice
   parameter, validated via reflect (must be non-nil pointer to slice).
2. NDJSON streaming logic moved into client.go::executeNDJSONRequest, which
   handles status code validation, Content-Type validation
   (application/x-ndjson or application/ndjson), and reflection-based
   streaming decode using json.NewDecoder. task_documents.go is now a thin
   wrapper that delegates to the helper.
3. Tests cover invalid dst (non-pointer, nil, non-slice), unexpected
   Content-Type, and the existing concatenated-NDJSON case.
